+++ include/Vector2.h
@@ -6,13 +6,13 @@ namespace pm {
 class Vector2
 {
 public:
-	double x;
-	double y;
+	float x;
+	float y;
 
 	Vector2() : x(0.0), y(0.0) { }
-	Vector2(double xx, double yy) : x(xx), y(yy) { }
+	Vector2(float xx, float yy) : x(xx), y(yy) { }
 
-	inline void set(double xx, double yy) { x = xx; y = yy; }
+	inline void set(float xx, float yy) { x = xx; y = yy; }
 
 	Vector2 operator-() const;
 
@@ -21,11 +21,11 @@ public:
 	Vector2 operator*(const Vector2 &v) const;
 	Vector2 operator/(const Vector2 &v) const;
 
-	Vector2 operator*(double s) const;
-	Vector2 operator/(double s) const;
+	Vector2 operator*(float s) const;
+	Vector2 operator/(float s) const;
 
-	friend Vector2 operator*(double s, const Vector2 &v);
-	friend Vector2 operator/(double s, const Vector2 &v);
+	friend Vector2 operator*(float s, const Vector2 &v);
+	friend Vector2 operator/(float s, const Vector2 &v);
 };
 
 inline Vector2 Vector2::operator-() const
@@ -53,22 +53,22 @@ inline Vector2 Vector2::operator/(const Vector2 &v) const
 	return Vector2(x / v.x, y / v.y);
 }
 
-inline Vector2 Vector2::operator*(double s) const
+inline Vector2 Vector2::operator*(float s) const
 {
 	return Vector2(x * s, y * s);
 }
 
-inline Vector2 Vector2::operator/(double s) const
+inline Vector2 Vector2::operator/(float s) const
 {
 	return Vector2(x / s, y / s);
 }
 
-inline Vector2 operator*(double s, const Vector2 &v)
+inline Vector2 operator*(float s, const Vector2 &v)
 {
 	return Vector2(s * v.x, s * v.y);
 }
 
-inline Vector2 operator/(double s, const Vector2 &v)
+inline Vector2 operator/(float s, const Vector2 &v)
 {
 	return Vector2(s / v.x, s / v.y);
 }
